third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/cycles-expected.txt

Test that cycles are not allowed in template content

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS template.content.appendChild(template) threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS template.content.appendChild(outerDiv) threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS innerDiv.appendChild(template) threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS innerDiv.appendChild(outerDiv) threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS innerTemplate.appendChild(outerDiv) threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS successfullyParsed is true

TEST COMPLETE
